rename handler to handlerclient
to avoid clashing with package names in common-fate/common-fate
