+++ pkg/handlerclient/local.go
@@ -1,4 +1,4 @@
-package handler
+package handlerclient
 
 import (
 	"context"
